feat: заготовка для тестов
