--- src/test/java/ru/yandex/practicum/eshop/it/eshopIt.java
@@ -1,56 +1,0 @@
-package ru.yandex.practicum.eshop.it;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-class eshopIt {
-
-  @BeforeEach
-  void setUp() {
-  }
-
-  @AfterEach
-  void tearDown() {
-  }
-
-  @Test
-  void redirectToMainItems() {
-  }
-
-  @Test
-  void getItems() {
-  }
-
-  @Test
-  void updateMainCartItems() {
-  }
-
-  @Test
-  void getCartItems() {
-  }
-
-  @Test
-  void updateCartItems() {
-  }
-
-  @Test
-  void getItemById() {
-  }
-
-  @Test
-  void updateItems() {
-  }
-
-  @Test
-  void buyItems() {
-  }
-
-  @Test
-  void getOrders() {
-  }
-
-  @Test
-  void getOrderById() {
-  }
-}
