+++ src/test/java/ru/yandex/practicum/eshop/utils/Data.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.eshop.utils;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class Data {
+
+}
