+++ src/test/resources/application.yml
@@ -1,0 +1,14 @@
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: update
+
+hibernate:
+  dialect: org.hibernate.dialect.H2Dialect
+  show_sql: true
+
+jdbc:
+  driverClassName: org.h2.Driver
+  url: jdbc:h2:mem:blog;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+  username: sa
+  password:
