+++ src/test/resources/schema.sql
@@ -1,0 +1,38 @@
+CREATE TABLE IF NOT EXISTS items(
+    id          BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    title       varchar,
+    image_path   varchar,
+    description varchar,
+    price double,
+    count       int
+);
+
+CREATE TABLE IF NOT EXISTS carts(
+    id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    total double
+);
+
+CREATE TABLE IF NOT EXISTS cart_item(
+     PRIMARY KEY (cart_id, item_id),
+     cart_id bigint,
+     item_id bigint,
+     count   int
+);
+
+CREATE TABLE IF NOT EXISTS orders(
+    id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    total_sum double
+);
+
+CREATE TABLE IF NOT EXISTS order_item(
+    PRIMARY KEY (order_id, item_id),
+    order_id bigint,
+    item_id bigint,
+    count   int
+);
+
+
+ALTER TABLE cart_item ADD FOREIGN KEY (cart_id) REFERENCES carts (id) ON DELETE CASCADE;
+ALTER TABLE cart_item ADD FOREIGN KEY (item_id) REFERENCES items (id) ON DELETE CASCADE;
+ALTER TABLE order_item ADD FOREIGN KEY (order_id) REFERENCES orders (id) ON DELETE CASCADE;
+ALTER TABLE order_item ADD FOREIGN KEY (item_id) REFERENCES items (id) ON DELETE CASCADE;
